Moved create_metadata_database_tables to common metadata manager
